Adicionando opção de confirmar delete

+++ resources/views/classeTerapeutica/index.blade.php
@@ -24,7 +24,7 @@
                                                 style="align-content: center"></i></a>
                                     </td>
                                     <td width="1%">
-                                        <a href="/classeTerapeutica.delete.{{ $obj->id }}"><i class="fa fa-trash"></i></a>
+                                        <a href="/classeTerapeutica.delete.{{ $obj->id }}" onclick="return confirm('Tem certeza de que deseja excluir este item?');"><i class="fa fa-trash"></i></a>
                                     </td>
                                     <td>
                                         {{ $obj->descricao }}
