+++ resources/views/item/index.blade.php
@@ -22,7 +22,7 @@
         <a href="/item.edit.{{ $obj->id }}"><i class="fa fa-edit" style="align-content: center;"></i></a>
       </td>
       <td width="1%">
-        <a href="/item.delete.{{ $obj->id }}"><i class="fa fa-trash"></i></a>
+        <a href="/item.delete.{{ $obj->id }}" onclick="return confirm('Tem certeza de que deseja excluir este item?');"><i class="fa fa-trash"></i></a>
       </td>
       <td>
         {{ $obj->descricao }}
